requirements.txt
mysql-connector-python
pyside6
pyside6-addons
tensorflow